# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateClientRequest:
    properties:
      address:
        type: string
      business_name:
        type: string
      email:
        type: string
      phone:
        type: string
      primary_contact:
        type: string
      tax_id:
        type: string
    required:
    - address
    - business_name
    - email
    - phone
    - primary_contact
    - tax_id
    type: object
...
  dto.CreateEquipmentRequest:
    properties:
      client_id:
        type: integer
      location:
        type: string
      model_id:
        type: integer
      serial_number:
        type: string
      status:
        type: string
    required:
    - client_id
    - location
    - model_id
    - serial_number
    - status
    type: object
...
  dto.UpdateClientRequest:
    properties:
      address:
        type: string
      business_name:
        type: string
      email:
        type: string
      phone:
        type: string
      primary_contact:
        type: string
      tax_id:
        type: string
    required:
    - address
    - business_name
    - email
    - phone
    - primary_contact
    - tax_id
    type: object
...
  dto.UpdateEquipmentRequest:
    properties:
      client_id:
        type: integer
      location:
        type: string
      model_id:
        type: integer
      serial_number:
        type: string
      status:
        type: string
    required:
    - client_id
    - location
    - model_id
    - serial_number
    - status
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@speckit.com
    name: API Support
  description: API for managing technical services, clients, equipment, work orders,
    and materials.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://swagger.io/terms/
  title: Service Management API
  version: 1.0.0
paths:
...
  /api/v1/clients:
    get:
      description: Retrieve a paginated list of clients with optional search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List clients
      tags:
      - Clients
    post:
      consumes:
      - application/json
      description: Register a new client
      parameters:
      - description: Client data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create client
      tags:
      - Clients
  /api/v1/clients/{id}:
    delete:
      description: Delete a client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete client
      tags:
      - Clients
    get:
      description: Retrieve a single client by their ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get client by ID
      tags:
      - Clients
    put:
      consumes:
      - application/json
      description: Update an existing client's data
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated client data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update client
      tags:
      - Clients
  /api/v1/equipment:
    get:
      description: Retrieve a paginated list of equipment with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by client ID
        in: query
        name: client_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List equipment
      tags:
      - Equipment
    post:
      consumes:
      - application/json
      description: Register new equipment for a client
      parameters:
      - description: Equipment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEquipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create equipment
      tags:
      - Equipment
  /api/v1/equipment/{id}:
    delete:
      description: Delete a piece of equipment by ID
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete equipment
      tags:
      - Equipment
    get:
      description: Retrieve a single piece of equipment by ID
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get equipment by ID
      tags:
      - Equipment
    put:
      consumes:
      - application/json
      description: Update an existing piece of equipment
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated equipment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update equipment
      tags:
      - Equipment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"